feat(api): move globals-usage route to scenarios router
Move the globals-usage endpoint from generator_configs router to the
scenarios router under /{name}/generators/{generator_name}/globals-usage,
and remove the now-unused Globals* models from generator_configs/models.py.
